Add MVP support for callback (#110)
This commit adds support for a callback in config. This allows mocking an API which will asynchronously make another API request as part of its contract.

A callback consists of a delay, a URL, a Content-Type and a Body. After the Delay, a POST request is made to the given URL, with the Content-Type + Body. The same substitutions are supported in the callback body as a response body, so eg. request values/fake data can be part of the API call made.

Limitations for callbacks in this MVP:

- Only POST is supported
- Custom headers aren't supported (only Content-Type is)
- Custom cookies aren't supported