Add support for callback (#110) (#111)
* Add MVP support for callback (#110)

This commit adds support for a callback in config. This allows mocking an API which will asynchronously make another API request as part of its contract.

A callback consists of a delay, a URL, a Content-Type and a Body. After the Delay, a POST request is made to the given URL, with the Content-Type + Body. The same substitutions are supported in the callback body as a response body, so eg. request values/fake data can be part of the API call made.

Limitations for callbacks in this MVP:

- Only POST is supported
- Custom headers aren't supported (only Content-Type is)
- Custom cookies aren't supported

* Add support for callback method and headers (#110)

- Use Go http.Client instead of http.Post
- Support any method
- Support arbitrary headers, instead of just Content-Type
- Improve logging, in particular where the statusCode from the callback is 4xx or 5xx

* Extract HandleCallback and add tests (#110)

- Move the callback body eval call to the dispatcher
- Move the rest into a new handle_callback.go
- Change to return resp, err to make it easier to test
- Add handle_callback_test.go

* Add Callback to README (#110)